# api/openapi.yaml
openapi: "3.1.0"
info:
  title: DirtOS REST API
  version: "1"
  description: |
    Local REST API embedded in the DirtOS desktop application.
    Intended for plugin development and 3rd-party integrations (n8n, Home Assistant, scripts, etc.).

    **Base URL:** `http://127.0.0.1:7272`
    Override the port with the `DIRTOS_API_PORT` environment variable before launching DirtOS.

    All endpoints return JSON. Timestamps are ISO 8601 strings in UTC.
    Enum values are lowercase snake_case strings unless noted otherwise.
servers:
  - url: http://127.0.0.1:7272
    description: DirtOS local API (default port)
...
paths:
...
  /api/v1/locations:
    get:
      tags: [Locations]
      summary: List locations in an environment
      operationId: listLocations
      parameters:
        - $ref: "#/components/parameters/environment_id"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      tags: [Locations]
      summary: Create a location
      operationId: createLocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewLocation"
      responses:
        "200":
          description: Created location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /api/v1/locations/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags: [Locations]
      summary: Get a location
      operationId: getLocation
      responses:
        "200":
          description: Location record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Locations]
      summary: Update a location
      operationId: updateLocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLocation"
      responses:
        "200":
          description: Updated location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Locations]
      summary: Delete a location
      operationId: deleteLocation
      responses:
        "204":
          description: Deleted successfully

  # ---------------------------------------------------------------------------
  # Species
  # ---------------------------------------------------------------------------
...
  /api/v1/plants:
    get:
      tags: [Plants]
      summary: List plants in an environment
      operationId: listPlants
      parameters:
        - $ref: "#/components/parameters/environment_id"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: List of plants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plant"
    post:
      tags: [Plants]
      summary: Create a plant
      operationId: createPlant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewPlant"
      responses:
        "200":
          description: Created plant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
  /api/v1/plants/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags: [Plants]
      summary: Get a plant
      operationId: getPlant
      responses:
        "200":
          description: Plant record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Plants]
      summary: Update a plant
      operationId: updatePlant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlant"
      responses:
        "200":
          description: Updated plant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Plants]
      summary: Delete a plant
      operationId: deletePlant
      responses:
        "204":
          description: Deleted successfully

  # ---------------------------------------------------------------------------
  # Schedules
  # ---------------------------------------------------------------------------
  /api/v1/schedules:
    get:
      tags: [Schedules]
      summary: List schedules in an environment
      operationId: listSchedules
      parameters:
        - $ref: "#/components/parameters/environment_id"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      tags: [Schedules]
      summary: Create a schedule
      operationId: createSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewSchedule"
      responses:
        "200":
          description: Created schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /api/v1/schedules/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags: [Schedules]
      summary: Get a schedule
      operationId: getSchedule
      responses:
        "200":
          description: Schedule record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Schedules]
      summary: Update a schedule
      operationId: updateSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSchedule"
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Schedules]
      summary: Delete a schedule
      operationId: deleteSchedule
      responses:
        "204":
          description: Deleted successfully

  # ---------------------------------------------------------------------------
  # Sensors
  # ---------------------------------------------------------------------------
  /api/v1/sensors:
    get:
      tags: [Sensors]
      summary: List sensors in an environment
      operationId: listSensors
      parameters:
        - $ref: "#/components/parameters/environment_id"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: List of sensors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sensor"
    post:
      tags: [Sensors]
      summary: Create a sensor
      operationId: createSensor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewSensor"
      responses:
        "200":
          description: Created sensor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
  /api/v1/sensors/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags: [Sensors]
      summary: Get a sensor
      operationId: getSensor
      responses:
        "200":
          description: Sensor record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Sensors]
      summary: Update a sensor
      operationId: updateSensor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSensor"
      responses:
        "200":
          description: Updated sensor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Sensors]
      summary: Delete a sensor
      operationId: deleteSensor
      responses:
        "204":
          description: Deleted successfully

  # ---------------------------------------------------------------------------
  # Issues
  # ---------------------------------------------------------------------------
...
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    environment_id:
      name: environment_id
      in: query
      required: true
      schema:
        type: integer
        format: int64
      description: Filter results to a specific environment
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 100
      description: Maximum number of records to return
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
      description: Number of records to skip (for pagination)
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: not found
  schemas:
...
    LocationType:
      type: string
      enum: [plot, space, tent, tray, pot, shed]
    Location:
      type: object
      properties:
        id:
          type: integer
          format: int64
        environment_id:
          type: integer
          format: int64
        parent_id:
          type: integer
          format: int64
          nullable: true
        location_type:
          $ref: "#/components/schemas/LocationType"
        name:
          type: string
        label:
          type: string
          nullable: true
        position_x:
          type: number
          nullable: true
        position_y:
          type: number
          nullable: true
        width:
          type: number
          nullable: true
        height:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        asset_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NewLocation:
      type: object
      required: [environment_id, location_type, name]
      properties:
        environment_id:
          type: integer
          format: int64
        parent_id:
          type: integer
          format: int64
          nullable: true
        location_type:
          $ref: "#/components/schemas/LocationType"
        name:
          type: string
        label:
          type: string
          nullable: true
        position_x:
          type: number
          nullable: true
        position_y:
          type: number
          nullable: true
        width:
          type: number
          nullable: true
        height:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
    UpdateLocation:
      type: object
      properties:
        parent_id:
          type: integer
          format: int64
          nullable: true
        location_type:
          $ref: "#/components/schemas/LocationType"
        name:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
        position_x:
          type: number
          nullable: true
        position_y:
          type: number
          nullable: true
        width:
          type: number
          nullable: true
        height:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true

    # -------------------------------------------------------------------------
    # Species
    # -------------------------------------------------------------------------
...
    PlantStatus:
      type: string
      enum: [planned, seedling, active, harvested, removed, dead]
    Plant:
      type: object
      properties:
        id:
          type: integer
          format: int64
        species_id:
          type: integer
          format: int64
          nullable: true
        location_id:
          type: integer
          format: int64
          nullable: true
        environment_id:
          type: integer
          format: int64
        status:
          $ref: "#/components/schemas/PlantStatus"
        name:
          type: string
        label:
          type: string
          nullable: true
        asset_id:
          type: string
          nullable: true
        planted_date:
          type: string
          format: date
          nullable: true
        germinated_date:
          type: string
          format: date
          nullable: true
        transplanted_date:
          type: string
          format: date
          nullable: true
        removed_date:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NewPlant:
      type: object
      required: [environment_id, name]
      properties:
        environment_id:
          type: integer
          format: int64
        species_id:
          type: integer
          format: int64
          nullable: true
        location_id:
          type: integer
          format: int64
          nullable: true
        status:
          $ref: "#/components/schemas/PlantStatus"
        name:
          type: string
        label:
          type: string
          nullable: true
        planted_date:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
    UpdatePlant:
      type: object
      properties:
        species_id:
          type: integer
          format: int64
          nullable: true
        location_id:
          type: integer
          format: int64
          nullable: true
        status:
          $ref: "#/components/schemas/PlantStatus"
        name:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
        planted_date:
          type: string
          format: date
          nullable: true
        germinated_date:
          type: string
          format: date
          nullable: true
        transplanted_date:
          type: string
          format: date
          nullable: true
        removed_date:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true

    # -------------------------------------------------------------------------
    # Schedule
    # -------------------------------------------------------------------------
    ScheduleType:
      type: string
      enum: [water, feed, maintenance, treatment, sample, custom]
    Schedule:
      type: object
      properties:
        id:
          type: integer
          format: int64
        environment_id:
          type: integer
          format: int64
          nullable: true
        plant_id:
          type: integer
          format: int64
          nullable: true
        location_id:
          type: integer
          format: int64
          nullable: true
        schedule_type:
          $ref: "#/components/schemas/ScheduleType"
        title:
          type: string
        cron_expression:
          type: string
          nullable: true
          description: Standard 5-field cron expression (e.g. "0 8 * * *" = 08:00 daily)
        next_run_at:
          type: string
          format: date-time
          nullable: true
        is_active:
          type: boolean
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NewSchedule:
      type: object
      required: [schedule_type, title]
      properties:
        environment_id:
          type: integer
          format: int64
          nullable: true
        plant_id:
          type: integer
          format: int64
          nullable: true
        location_id:
          type: integer
          format: int64
          nullable: true
        schedule_type:
          $ref: "#/components/schemas/ScheduleType"
        title:
          type: string
        cron_expression:
          type: string
          nullable: true
        is_active:
          type: boolean
          default: true
        notes:
          type: string
          nullable: true
    UpdateSchedule:
      type: object
      properties:
        schedule_type:
          $ref: "#/components/schemas/ScheduleType"
        title:
          type: string
          nullable: true
        cron_expression:
          type: string
          nullable: true
        is_active:
          type: boolean
          nullable: true
        plant_id:
          type: integer
          format: int64
          nullable: true
        location_id:
          type: integer
          format: int64
          nullable: true
        notes:
          type: string
          nullable: true

    # -------------------------------------------------------------------------
    # Sensor
    # -------------------------------------------------------------------------
    SensorType:
      type: string
      enum: [moisture, light, temperature, humidity, ph, ec, co2, air_quality, custom]
    SensorConnectionType:
      type: string
      enum: [serial, usb, mqtt, http, manual, home_assistant]
    Sensor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        environment_id:
          type: integer
          format: int64
          nullable: true
        location_id:
          type: integer
          format: int64
          nullable: true
        plant_id:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        sensor_type:
          $ref: "#/components/schemas/SensorType"
        connection_type:
          $ref: "#/components/schemas/SensorConnectionType"
        connection_config_json:
          type: string
          nullable: true
          description: JSON blob with connection-specific configuration
        poll_interval_seconds:
          type: integer
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NewSensor:
      type: object
      required: [name, sensor_type, connection_type]
      properties:
        environment_id:
          type: integer
          format: int64
          nullable: true
        location_id:
          type: integer
          format: int64
          nullable: true
        plant_id:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        sensor_type:
          $ref: "#/components/schemas/SensorType"
        connection_type:
          $ref: "#/components/schemas/SensorConnectionType"
        connection_config_json:
          type: string
          nullable: true
        poll_interval_seconds:
          type: integer
          nullable: true
        is_active:
          type: boolean
          default: true
    UpdateSensor:
      type: object
      properties:
        name:
          type: string
          nullable: true
        sensor_type:
          $ref: "#/components/schemas/SensorType"
        connection_type:
          $ref: "#/components/schemas/SensorConnectionType"
        connection_config_json:
          type: string
          nullable: true
        poll_interval_seconds:
          type: integer
          nullable: true
        location_id:
          type: integer
          format: int64
          nullable: true
        plant_id:
          type: integer
          format: int64
          nullable: true
        is_active:
          type: boolean
          nullable: true

    # -------------------------------------------------------------------------
    # Issue
    # -------------------------------------------------------------------------
...